backend/tools/requirements_dev.txt
sqlalchemy-schemadisplay==1.3
pydot==1.4.2
graphviz==0.20.1
eralchemy2==1.3.6 